casting

+++ proj1/ngordnet/NGramMap.java
@@ -106,7 +106,7 @@ public class NGramMap {
         TimeSeries<Integer> count = countHistory(word, startYear, endYear);
         TimeSeries<Double> time = new TimeSeries<Double>();
         for (int year = startYear; year <= endYear; year++) {
-            if (timeMap.containsKey(year) && count.containsKey(year)) {
+            if (timeMap.containsKey((long) year) && count.containsKey(year)) {
                 time.put(year, timeMap.get(year).doubleValue());
             }
         }
@@ -119,7 +119,7 @@ public class NGramMap {
         TimeSeries<Integer> count = countHistory(word);
         TimeSeries<Double> time = new TimeSeries<Double>();
         for (long year : timeMap.keySet()) {
-            if (timeMap.containsKey(year) && count.containsKey(year)) {
+            if (timeMap.containsKey(year) && count.containsKey((int) year)) {
                 time.put((int) year, timeMap.get((int) year).doubleValue());
             }
         }
